refactor: extract the stat tile grid and month calendar
Both the Insights screen and the per-book stats screen that follows show the
same 2x2 tile grid and the same month calendar with different copy behind
them. Copying 240 lines of render() is how the second screen drifts from the
first, so pull them out first.

The code moves verbatim -- Insights renders identically.

Data reaches the calendar through a MonthSource function-pointer pair rather
than a std::function or a virtual: neither store shares a base class, and this
is called from render().
